Move repr/str tests to their corresponding test files
- test_fcpath_repr_str → tests/test_file_cache_path.py
- test_filecachesource_repr_str → tests/test_file_cache_source.py
- test_filecache_repr_str stays in tests/test_file_cache.py
