+++ tests/test_file_cache.py
@@ -3163,54 +3163,3 @@ def test_filecache_repr_str():
         assert 'anonymous=True' in r
         assert 'lock_timeout=30' in r
         assert 'nthreads=4' in r
-
-
-def test_fcpath_repr_str():
-    # Default FCPath — only the path
-    p = FCPath('gs://bucket/path/to/file')
-    assert repr(p) == "FCPath('gs://bucket/path/to/file')"
-    assert str(p) == 'gs://bucket/path/to/file'
-
-    # FCPath with extra options — they appear in repr
-    p2 = FCPath('s3://bucket/path', anonymous=True, lock_timeout=10, nthreads=2)
-    r = repr(p2)
-    assert r.startswith('FCPath(')
-    assert 'anonymous=True' in r
-    assert 'lock_timeout=10' in r
-    assert 'nthreads=2' in r
-
-    # filecache kwarg appears in repr when set
-    with FileCache(cache_name=None, delete_on_exit=True) as fc:
-        p3 = FCPath('gs://bucket/x', filecache=fc)
-        r3 = repr(p3)
-        assert 'filecache=' in r3
-
-
-def test_filecachesource_repr_str():
-    from filecache import FileCacheSourceGS, FileCacheSourceS3, FileCacheSourceFake
-    from filecache import FileCacheSourceFile, FileCacheSourceHTTP
-
-    src_gs = FileCacheSourceGS('gs', 'rms-filecache-tests', anonymous=True)
-    r = repr(src_gs)
-    assert r == "FileCacheSourceGS('gs', 'rms-filecache-tests', anonymous=True)"
-    assert str(src_gs) == 'gs://rms-filecache-tests'
-
-    src_s3 = FileCacheSourceS3('s3', 'rms-filecache-tests', anonymous=True)
-    r = repr(src_s3)
-    assert r == "FileCacheSourceS3('s3', 'rms-filecache-tests', anonymous=True)"
-    assert str(src_s3) == 's3://rms-filecache-tests'
-
-    src_http = FileCacheSourceHTTP('https', 'storage.googleapis.com', anonymous=False)
-    r = repr(src_http)
-    assert r == "FileCacheSourceHTTP('https', 'storage.googleapis.com', anonymous=False)"
-    assert str(src_http) == 'https://storage.googleapis.com'
-
-    src_file = FileCacheSourceFile('file', '', anonymous=False)
-    r = repr(src_file)
-    assert r == "FileCacheSourceFile('file', '', anonymous=False)"
-    assert str(src_file) == 'file://'
-
-    src_fake = FileCacheSourceFake('fake', 'fake-bucket', anonymous=False)
-    r = repr(src_fake)
-    assert r == "FileCacheSourceFake('fake', 'fake-bucket', anonymous=False)"
-    assert str(src_fake) == 'fake://fake-bucket'
