+++ tests/test_file_cache_path.py
@@ -1308,3 +1308,24 @@ def test_splitpath():
                                                               FCPath('d'))
     assert FCPath('gs://bucket/a/c/b/b1/c/d/d1').splitpath('c') == \
         (FCPath('gs://bucket/a'), FCPath('b/b1'), FCPath('d/d1'))
+
+
+def test_fcpath_repr_str():
+    # Default FCPath — only the path
+    p = FCPath('gs://bucket/path/to/file')
+    assert repr(p) == "FCPath('gs://bucket/path/to/file')"
+    assert str(p) == 'gs://bucket/path/to/file'
+
+    # FCPath with extra options — they appear in repr
+    p2 = FCPath('s3://bucket/path', anonymous=True, lock_timeout=10, nthreads=2)
+    r = repr(p2)
+    assert r.startswith('FCPath(')
+    assert 'anonymous=True' in r
+    assert 'lock_timeout=10' in r
+    assert 'nthreads=2' in r
+
+    # filecache kwarg appears in repr when set
+    with FileCache(cache_name=None, delete_on_exit=True) as fc:
+        p3 = FCPath('gs://bucket/x', filecache=fc)
+        r3 = repr(p3)
+        assert 'filecache=' in r3
